requirements/base.txt
torch>=2.5.0
torchvision>=0.20.0
transformers>=4.50.0
peft>=0.13.0
trl>=0.12.0
accelerate>=1.0.0
datasets>=3.0.0
pillow>=10.0.0
pydantic>=2.0.0
pyyaml>=6.0
sacrebleu>=2.4.0
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.9
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
tensorboard>=2.18.0
pyahocorasick>=2.1.0
qwen-vl-utils>=0.0.10
safetensors>=0.4.0
imagehash>=4.3.0
